still fixing map checking..

+++ draw.c
@@ -6,7 +6,7 @@
 /*   By: abchikhi <abchikhi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/22 18:16:05 by abchikhi          #+#    #+#             */
-/*   Updated: 2024/01/13 00:46:19 by abchikhi         ###   ########.fr       */
+/*   Updated: 2024/01/13 03:49:33 by abchikhi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ void	change_pixel_color(t_data *img, int x, int y, int change)
 	if (x < 0 || x >= WIDTH || y < 0 || y >= HEIGHT)
 		return ;
 	dst = img->addr + (y * img->line_length + x * (img->bits_per_pixel / 8));
-    if ((*(unsigned int*)dst + change) < 0)
+    if ((unsigned)(*(unsigned int*)dst + change) < 0)
         return ;
 	*(unsigned int*)dst += change;
 }
